Dashboard widgets: filter out Default repo in queries if app config has 'settings.hideDefaultProject' #5149
